enact cruel censorship
